Working Encrypt, Decrypt, & Cracker

+++ lib/decrypt.rb
@@ -18,12 +18,26 @@ class Decrypt
     decrypted_message = StringRotator.new(contents, @offset).decrypt
 
     file_name = file_to_write
-    decrypted_file = File.open(file_name, "w")
-    decrypted_file.puts decrypted_message
-    decrypted_file.close
 
-    @printer.puts "Created '#{file_to_write}' with the key #{@key} and date #{@date}"
+    if File.exists?(file_to_write)
+      puts "A file by the name of '#{file_to_write}' already exists. Are you sure you want to overwrite it? Yes/No"
+      input = $stdin.gets.chomp
+      if input.downcase == "yes"
+        decrypted_file = File.open(file_name, "w")
+        decrypted_file.puts decrypted_message
+        decrypted_file.close
+        @printer.puts "Created '#{file_to_write}' with the key #{@key} and date #{@date}"
+      else
+        abort("Program canceled.")
+      end
+    else
+      decrypted_file = File.open(file_name, "w")
+      decrypted_file.puts decrypted_message
+      decrypted_file.close
+      @printer.puts "Created '#{file_to_write}' with the key #{@key} and date #{@date}"
+    end
   end
+
 
   def file_to_read
     ARGV[0]
