+++ lib/encrypt.rb
@@ -18,11 +18,24 @@ class Encrypt
     encrypted_message = StringRotator.new(contents, @offset).encrypt
 
     file_name = file_to_write
-    encrypted_file = File.open(file_name, "w")
-    encrypted_file.puts encrypted_message
-    encrypted_file.close
 
-    @printer.puts "Created '#{file_to_write}' with the key #{@key.key} and date #{@date.date}"
+    if File.exists?(file_to_write)
+      puts "A file by the name of '#{file_to_write}' already exists. Are you sure you want to overwrite it? Yes/No"
+      input = $stdin.gets.chomp
+        if input.downcase == "yes"
+          encrypted_file = File.open(file_name, "w")
+          encrypted_file.puts encrypted_message
+          encrypted_file.close
+          @printer.puts "Created '#{file_to_write}' with the key #{@key.key} and date #{@date.date}"
+        else
+          abort("Program canceled.")
+        end
+    else
+      encrypted_file = File.open(file_name, "w")
+      encrypted_file.puts encrypted_message
+      encrypted_file.close
+      @printer.puts "Created '#{file_to_write}' with the key #{@key.key} and date #{@date.date}"
+    end
   end
 
   def file_to_read
